Add logic to relaunch range orders
The idea behind range orders is to relaunch them after completion of a
succesfful transaction related to the range order.

The new upper limit of the range order will be computed with the
following formula:
`new_max_amount = max_amount - selected_amount`

Range orders will be relaunched only if the `new_max_amount` is above
the min amount provided in the original range order.